requirements.txt
Flask==2.2.3
python-dotenv==0.21.0
langchain==0.0.92
ollama==0.1.0
PyPDF2==3.0.1
python-docx==0.8.11